requirements.txt
sounddevice
numpy
SpeechRecognition
pyttsx3
anthropic
tavily-python
python-dotenv
tzdata
yt-dlp
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
